Turn backwards tail recursion into forward non-tail recursion
This makes the function a little terser and removes the need for an
accumulator, but otherwise doesn't seem to make a difference.

I tested this by duplicating `input` to 32k lines (1.4MB file), and
running both versions on it. With `time`, I could not see any
significant difference in time or memory consumption.